remove or move table-related types to their new homes
