+++ src/frontend/src/views/Data/transforms.tsx
@@ -1,9 +1,14 @@
 import { generateOrgSpecificUrl, ORG_API } from "src/common/api";
 import ENV from "src/common/constants/ENV";
 import type { TreeType } from "src/common/constants/types";
-import { Transform } from "src/common/types/data";
 
 const { API_URL } = ENV;
+
+interface Transform {
+  key: string;
+  inputs: string[];
+  method: (inputs: any[]) => any;
+}
 
 //* (mlila): below, we want to use the tree id to generate the urls in question.
 //* Links can only be generated for workflows with a tree
